% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{sigverse_conflicts}
\alias{sigverse_conflicts}
\title{Conflicts between the sigverse and other packages}
\usage{
sigverse_conflicts(only = NULL)
}
\arguments{
\item{only}{Set this to a character vector to restrict to conflicts only
with these packages.}
}
\description{
This function lists all the conflicts between packages in the sigverse
and other packages that you have loaded.
}
\examples{
sigverse_conflicts()
}
