% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{sigverse_deps}
\alias{sigverse_deps}
\title{List all sigverse dependencies}
\usage{
sigverse_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
sigverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
List all sigverse dependencies
}
