% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sigverse_packages}
\alias{sigverse_packages}
\title{List all packages in the sigverse}
\usage{
sigverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include sigverse in the list?}
}
\description{
List all packages in the sigverse
}
\examples{
sigverse_packages()
}
